coverity fix - result is not floating point - MapGrid2D.cpp

+++ src/libYARP_dev/src/MapGrid2D.cpp
@@ -342,7 +342,7 @@ bool MapGrid2D::loadMapYarpAndRos(string yarp_filename, string ros_yaml_filename
             for (size_t x = 0; x < m_width; x++)
             {
                 yarp::sig::PixelRgb pix_occ = ros_img.safePixel(x, y);
-                double color_avg = (pix_occ.r + pix_occ.g + pix_occ.b) / 3;
+                int color_avg = (pix_occ.r + pix_occ.g + pix_occ.b) / 3;
                 unsigned char occ = (unsigned char)((255 - color_avg) / 255.0);
                 m_map_occupancy.safePixel(x, y) = occ * 100;
             }
@@ -391,7 +391,7 @@ bool MapGrid2D::loadMapROSOnly(string ros_yaml_filename)
         for (size_t x = 0; x < m_width; x++)
         {
             yarp::sig::PixelRgb pix_occ = ros_img.safePixel(x, y);
-            double color_avg = (pix_occ.r + pix_occ.g + pix_occ.b) / 3;
+            int color_avg = (pix_occ.r + pix_occ.g + pix_occ.b) / 3;
             unsigned char occ = (unsigned char)((255 - color_avg) / 255.0);
             m_map_occupancy.safePixel(x, y) = occ * 100;
         }
